refactor: site -> study area
