+++ src/service/index.js
@@ -3,6 +3,6 @@ export * from './modules/annotations';
 export * from './modules/config';
 export * from './modules/notifications';
 export * from './modules/projects';
-export * from './modules/site';
+export * from './modules/studyAreas';
 export * from './modules/upload';
 export * from './modules/dataFields';
