+++ src/store/modules/projects/studyAreas.js
